fix: change import and rename to local
